requirements.txt
requests
sqlalchemy
progressbar2
